defer all spawning to user code
